requirement.txt
streamlit
streamlit
ultralytics
torch
torchvision
Pillow
numpy